+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 - Added possibility to highlight mouse target cell on the schedule (can be enabled on settings)
 - Added subject code to paco schedule scraper and to the extension itself
 - Added mechanism to load each subject schedule with all possible classes (with message logging each load for the user)
+- Right click on a class cell opens context menu with the following options:
+    - Complete Schedule: show all classes for that subject
+    - Edit Class: edit class information
+    - Remove Class: remove that class from the schedule
+    - Remove Subject: remove all classes from that subject from the schedule
+- Clicking on an empty cell in the schedule allows to add a new class
 
 ---